fix + refactor: Fixed and refactored upgrade premium flow
