+++ nimbus-web/components/Layout.tsx
@@ -3,8 +3,12 @@ import React from "react";
 import Navbar from "./Navbar/Navbar";
 import Loading from "./Loading";
 import Footer from "./Footer";
-import checkPremiumOnNewSession from "@/utils/checkPremiumOnNewSession";
+import updateOnNewSession from "@/utils/updateOnNewSession";
 import { useRouter } from "next/router";
+import getPremiumExpire from "@/utils/getPremiumExpire";
+import getPremiumType from "@/utils/getPremiumType";
+import checkPremiumExpire from "@/utils/checkPremiumExpire";
+import checkPremium from "@/utils/checkPremium";
 
 const excludePathnameList = ["/plan", "/map"];
 
@@ -17,7 +21,12 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
     React.useEffect(() => {
         if (user) {
             console.log("check premium status on new session");
-            checkPremiumOnNewSession(user, router);
+            updateOnNewSession(user, router);
+            const metadata = {
+                premium_type: getPremiumType(user),
+                premium_expire: getPremiumExpire(user, "number"),
+            };
+            checkPremium(metadata, user, router);
         }
     }, [user, router]);
     return !isLoading ? (
